se ajusta el costo total en COP de la operacion para el historial del usuario

+++ lib/services/models/history_operations_user/response/user.dart
@@ -16,7 +16,7 @@ class User{
       _$UserFromJson(json);
 
  String nickName;
- int rateSeller;
+ double rateSeller;
  double rateBuyer;
  String address;
